Refactor main entry code (skeleton)
