+++ setup.cfg
@@ -5,7 +5,7 @@
 
 [metadata]
 name = empire
-description = Add a short description here!
+description = Game about empires
 author = Erez-Shlingbaum, segalmatan
 author_email = erez.shlingbaum@gmail.com
 license = MIT
